initial  commit, myparser duplications  removed, new subpackages created for better organization
