+++ genome_files/get_fagz_sizes.py
@@ -11,12 +11,7 @@ import shutil
 import gc
 import re
 
-
-class MyParser(argparse.ArgumentParser):
-    def error(self, message):
-        sys.stderr.write("error: %s\n" % message)
-        self.print_help()
-        sys.exit(2)
+from protoSpaceJAM.util.utils import MyParser
 
 
 def parse_args():
