+++ main.py
@@ -1,13 +1,8 @@
 import os.path
-from scripts.utils import *
+from protoSpaceJAM.util.utils import *
 import traceback
 import time
 
-class MyParser(argparse.ArgumentParser):
-    def error(self, message):
-        sys.stderr.write("error: %s\n" % message)
-        self.print_help()
-        sys.exit(2)
 
 def parse_args():
     parser = MyParser(description="protoSpaceJAM")
