+++ protoSpaceJAM/precompute/gRNA_scan_split4mapping.py
@@ -12,16 +12,11 @@ import gc
 import math
 import errno, os, stat, shutil
 
-sys.path.insert(1, "..")
+from protoSpaceJAM.util.utils import MyParser
+
+sys.path.insert(1, "../..")
 from gRNA_search import *
 from utils import *
-
-
-class MyParser(argparse.ArgumentParser):
-    def error(self, message):
-        sys.stderr.write("error: %s\n" % message)
-        self.print_help()
-        sys.exit(2)
 
 
 def parse_args():
