+++ protoSpaceJAM/serialize_fa.py
@@ -12,12 +12,7 @@ import gc
 import re
 import pickle
 
-
-class MyParser(argparse.ArgumentParser):
-    def error(self, message):
-        sys.stderr.write("error: %s\n" % message)
-        self.print_help()
-        sys.exit(2)
+from protoSpaceJAM.util.utils import MyParser
 
 
 def parse_args():
